requirements.txt
paho-mqtt>=2.1.0
pydantic>=2.8.0
pyyaml>=6.0.0
rich>=13.7.0
